[fixed] small fixes
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1655 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/packets/scene/UpdateTransformWithParentMessage.h
@@ -68,7 +68,7 @@ public:
 		insertByte(0); // unknown
 		
 		// add direction
-		insertByte(object->getDirectionAngle());
+		insertByte((byte) object->getDirectionAngle());
 		
 		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 			 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << "," 
@@ -94,7 +94,7 @@ public:
 		insertByte(0); // unknown
 
 		// add direction
-		insertByte(object->getDirectionAngle());
+		insertByte((byte) object->getDirectionAngle());
 
 		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 				 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << ","
